Added restrictions so regular members cannot use moderation commands

+++ src/Commands/Moderation/dm.ts
@@ -6,27 +6,28 @@ export const command: Command = {
     aliases: [],
     run: async(client, msg, args) => {
         if (args.length < 2) return;
+        if (!msg.guild.member(msg.author).roles.cache.has('643128678804619316')) return;
         const message = args.shift().toString().replace(/,/g, ' ');
 
-    msg.channel.delete()
-    .then(() => {
-        msg.mentions.members.first().send(message)
-        .then((messag) => {
-            msg.channel.send("Done!")
-            .then((m) => {
-                m.delete({timeout: 4000});
+        msg.channel.delete()
+        .then(() => {
+            msg.mentions.members.first().send(message)
+            .then((messag) => {
+                msg.channel.send("Done!")
+                .then((m) => {
+                    m.delete({timeout: 4000});
+                })
             })
-        })
-        .catch((err) => {
-            msg.channel.send("It seems like the person you're trying to DM has their DMs closed :c")
-            .then((m) => {
-                m.delete({timeout: 4000});
-            })
-            .catch((e) => {
-                console.error(e);
+            .catch((err) => {
+                msg.channel.send("It seems like the person you're trying to DM has their DMs closed :c")
+                .then((m) => {
+                    m.delete({timeout: 4000});
+                })
+                .catch((e) => {
+                    console.error(e);
+                });
             });
         });
-    });
 
     }
 }
